+++ custom.d.ts
@@ -1,7 +1,8 @@
 declare var OC: {
     generateUrl(string, any?): string,
     PasswordConfirmation: any,
-    getCurrentUser(): {uid: string, displayName: string}
+    getCurrentUser(): {uid: string, displayName: string},
+    requestToken: string,
 };
 declare var oc_requesttoken: string;
 declare var JSXC: any;