+++ tests/bootstrap-integration.php
@@ -3,6 +3,8 @@ $nc_root = __DIR__ . '/../../..';
 
 @include 'bootstrap-config.development.php';
 
+echo "Using ".realpath($nc_root)." as Nextcloud root.\n\n";
+
 try {
 	require_once $nc_root . '/tests/bootstrap.php';
 } catch (Exception $ex) {