+++ webpack.config.js
@@ -3,7 +3,7 @@ const path = require("path");
 const webpack = require("webpack");
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
+const {CleanWebpackPlugin} = require('clean-webpack-plugin');
 
 const DESTINATION_DIR = 'dist/';
 
@@ -102,9 +102,9 @@ const config = {
       new MiniCssExtractPlugin({
          filename: 'css/bundle.css',
       }),
-      new CleanWebpackPlugin([DESTINATION_DIR]),
+      new CleanWebpackPlugin(),
       new CopyWebpackPlugin([{
-         from: 'node_modules/jsxc/dist/',
+         from: 'node_modules/@jsxc/jsxc/dist/',
          to: 'js/jsxc/'
       }, {
          from: 'node_modules/libsignal-protocol/dist/',
@@ -112,6 +112,9 @@ const config = {
       }, {
          from: 'appinfo/',
          to: 'appinfo/'
+      }, {
+         from: 'js/',
+         to: 'js/'
       }, {
             from: 'img/',
             to: 'img/'