+++ CHANGELOG.md
@@ -4,10 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased 4.0]
+## [Unreleased]
+
+## 4.0.0 - 2020-04-08
+### Fixed
+- default stun server
+
 ### Changed
 - update JSXC to v4.0.0 ([change log](https://github.com/jsxc/jsxc/blob/refactoring/CHANGELOG.md))
-- bump min NC version to 14
+- bump min NC version to 16
+- bump max NC version to 18
+- use db migration
+- show always login with and without chat buttons
+- preserve local storage during logout
+- adapt feature policy
 
 ## 3.4.4 - 2019-05-10
 ### Fixed